added new /matches route template
